pyproject.toml
[project]
name = "weaviate-agents-workshop"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "datasets>=3.6.0",
    "jupyter>=1.1.1",
    "python-dotenv>=1.1.0",
    "tqdm>=4.67.1",
    "weaviate-agents>=0.8.2",
    "weaviate-client>=4.15.0",
]
